backend/requirements.txt

fastapi
uvicorn[standard]
sqlmodel
python-jose[cryptography]
bcrypt
python-multipart
psycopg2-binary
python-dotenv
pytz
pydantic-settings
